+++ src/app/Scenes/Onboarding/OnboardingStore.tsx
@@ -1,0 +1,18 @@
+import { createContextStore } from "easy-peasy"
+
+export interface OnboardingStoreModel {
+  step:
+    | "WelcomeStep"
+    | "EmailStep"
+    | "LoginPasswordStep"
+    | "SignUpPasswordStep"
+    | "NameStep"
+    | "ForgotPasswordStep"
+}
+
+export const OnboardingStore = createContextStore<OnboardingStoreModel>((runtimeModel) => {
+  return {
+    step: "WelcomeStep",
+    ...runtimeModel,
+  }
+})
